fixes text partagé

+++ src/main/java/com/example/Controller/NetworkController.java
@@ -44,7 +44,7 @@ public class NetworkController {
 		//this.ctrl.getChatController().onMessageReceived(res);
 	}
 
-	public byte[] IntToByte(int value) {
+	public byte[] IntToByte(short value) {
 		return this.networkModel.IntToByteArray(value);
 	}
 	public byte[] concatenateByteArrays(byte[] array1, byte[] array2) {
